rename: 파일명 snake_case로 변경
lower camel case에서 플러터 관습인 snake_case로 변경

TT-78 scriptInput
