+++ lib/features/scriptInput/viewModel/state/script_input_view_model_state.dart
